refactor: change name folder model to conection and remake all path refer connection
